requirements.txt
asgiref==3.5.0
astroid==2.12.13
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
coverage==6.3.2
cryptography==39.0.0
defusedxml==0.7.1
dill==0.3.6
Django==4.0.2
django-crontab==0.7.1
django-filter==21.1
django-phonenumber-field==7.0.2
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
djoser==2.1.0
drf-yasg==1.21.4
future==0.18.2
idna==3.4
importlib-metadata==4.11.1
inflection==0.5.1
isort==5.11.3
itypes==1.2.0
Jinja2==3.1.2
lazy-object-proxy==1.8.0
Markdown==3.3.6
MarkupSafe==2.1.1
mccabe==0.7.0
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==22.0
phonenumbers==8.13.4
platformdirs==2.6.0
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.3.0
pylint==2.15.9
python3-openid==3.2.0
pytz==2021.3
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
serializers==0.2.4
simplejson==3.18.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
tomlkit==0.11.6
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
wrapt==1.14.1
zipp==3.7.0